Extract nameToInitials to util
